+++ HornyGCC/Requester.h
@@ -4,7 +4,7 @@ void ErstelleAslReqs(void);
 void EntferneAslReqs(void);
 BOOL AslProjLaden(void);
 BOOL AslProjSpeichern(void);
-LONG Frage(STRPTR text, STRPTR knopf);
+int32 Frage(STRPTR text, STRPTR knopf);
 void Meldung(STRPTR text);
 BOOL AslSMFLaden(void);
 BOOL AslSMFSpeichern(void);
